Added returnvalue to function Sleep in aplibs/timers.cc and typecast in fastlst/response.cc. This removes warnings under gcc 2.95 and (hopefully :-) allows compilation under gcc 3.2.

+++ fastlst/response.cc
@@ -223,7 +223,7 @@ void SetOrigin (char *buffer, char *origin, ADR *adr)
     char *c;
 
     c = addrs (adr);            // max 79chars in origin
-    sprintf (buffer, " * Origin: %.*s (%s)\r", 65 - strlen (c), origin, c);
+    sprintf (buffer, " * Origin: %.*s (%s)\r", 65 - (int)strlen(c), origin, c);
 }
